+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ## [11.0.3] (IN PROGRESS)
 
 * Fetch facets with the same query as `Contributor` records when the search is done by selecting browse result. Refs UIIN-2859.
+* Change the `Effective location` call numbers facet query to not see locations assigned to items with an empty `Effective call number` field. Refs UIIN-2871.
 
 ## [11.0.2](https://github.com/folio-org/ui-inventory/tree/v11.0.2) (2024-04-19)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.1...v11.0.2)